refactor orchestrator logsagentexporter
